data_analyze_project/requirements.txt
asgiref==3.8.1
confluent-kafka==2.6.0
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
Django==5.1.2
djangorestframework==3.15.2
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
patsy==0.5.6
pandas==2.2.3
pika==1.3.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scipy==1.14.1
six==1.16.0
statsmodels==0.14.4
sqlparse==0.5.1
tzdata==2024.2
requests==2.32.3
reportlab==4.2.5
pillow==10.4.0
